Added material colors to about song sheet. Refactoring
