+++ composeApp/src/commonMain/composeResources/values/strings.xml
@@ -277,6 +277,7 @@
 	<string name="action_refresh">Refresh</string>
 	<string name="action_share">Share</string>
 	<string name="action_track_info">View track info</string>
+	<string name="track_info_title">Track info</string>
 	<string name="action_view_album">View album</string>
 	<string name="action_view_playlist">View playlist</string>
 	<string name="action_view_artist">View artist</string>
