+++ composeApp/src/commonMain/kotlin/paige/navic/ui/navigation/Screen.kt
@@ -100,7 +100,7 @@ sealed interface Screen : NavKey {
 
 	@Immutable
 	@Serializable
-	data class SongDetail(val songId: String) : Screen
+	data class SongDetail(val songId: String, val coverArtId: String? = null) : Screen
 
 	@Immutable
 	@Serializable
