+++ composeApp/src/commonMain/kotlin/paige/navic/ui/screens/nowPlaying/components/controls/MoreButton.kt
@@ -89,7 +89,7 @@ fun NowPlayingMoreButton(
 				},
 				onTrackInfo = dropUnlessResumed {
 					expanded = false
-					backStack.add(Screen.SongDetail(song.id))
+					backStack.add(Screen.SongDetail(songId = song.id, coverArtId = song.coverArtId))
 				},
 				rating = songRating,
 				onSetRating = onSetSongRating,
